+++ NEWS.md
@@ -1,3 +1,13 @@
+# c14bazAAR 1.2.0
+
+## general changes
+- unified database names in all functions, tables, variables and documentation (#86)
+
+## new features
+- added a basic plot function for c14_date_lists (#82)
+- added a basic write function for c14_date_lists: `write_c14()` (#84)
+- added a version column that documents from which database version a certain date is pulled (#85)
+
 # c14bazAAR 1.1.0
 
 ## general changes
@@ -5,13 +15,14 @@
 - moved main development repository to github/ropensci/c14bazAAR (e0e6827f0381be04c50380eec277c01cad44ac7d)
 - created [c14bazAAR project](https://doi.org/10.17605/OSF.IO/3DS6A) on the OSF platform with a DOI
 - more work on an article for the Journal of Open Source Software (paper.md + paper.bib)
+- changed citation to JOSS article after its publication
 
 ## new features
 - new download interface as suggested by Enrico Crema in the ROpenSci review: `get_c14data()` (#76)
 - replaced hard coded URLs with arguments to get helper functions (caabcb7b)
 
 ## new getter functions
-- added getter function for IRDD database: `get_IRDD` (#79)
+- added getter function for irdd database: `get_irdd` (#79)
 
 # c14bazAAR 1.0.3
 
@@ -24,8 +35,8 @@
 - created a completely artificial example dataset that replaces the sampled version
 
 ## new getter functions
-- added getter function for Palmisano et al. database: `get_Palmisano` (#59)
-- added getter function for EUBAR database: `get_EUBAR` (#64)
+- added getter function for Palmisano et al. database: `get_palmisano` (#59)
+- added getter function for eubar database: `get_eubar` (#64)
 
 ## new features
 - added new options for the deduplication function (see `?duplicates`) (#63)
@@ -33,7 +44,7 @@
 - better implementation of the `c14_date_list` as a subclass of tibble for a better integration of the subclass into the tidyverse (#67)
 
 ## bugfixes
-- small file path construction fix in CONTEXT getter function
+- small file path construction fix in context getter function
 - replaced some deprecated functions by other packages (dplyr::funs & tibble::as.tibble)
 - replaced `RCurl::url.exists` with `httr::http_error` in `check_connection_to_url` (#68)
 - fixed `as.sf` error that occurred when date lists with dates without coordinates were transformed